backend-api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
pillow==10.0.1
numpy==1.24.3
tensorflow==2.13.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
werkzeug==2.3.7
PyJWT==2.8.0
